requirements.txt
Flask==0.10.1
Flask-Assets==0.10
Flask-WeasyPrint==0.5
flask-wtf==0.9.3
itsdangerous==0.24
Jinja2==2.7.3
libsass==0.8.2
MarkupSafe==0.23
mypy-lang==0.2.0
python-dateutil==2.4
WeasyPrint==0.23
Werkzeug==0.9.6
wtforms==1.0.5
